requirements_vulnerable.txt
# Vulnerable dependencies - will trigger GitHub Security alerts
Flask==1.0.2          # CVE-2019-1010083 - Path traversal vulnerability
requests==2.19.0      # CVE-2018-18074 - SSL verification bypass  
Jinja2==2.8.1         # CVE-2019-10906 - Sandbox escape
SQLAlchemy==1.2.0     # CVE-2019-7164 - SQL injection via order_by
PyYAML==3.13          # CVE-2017-18342 - Arbitrary code execution
Pillow==5.2.0         # CVE-2019-16865 - Buffer overflow
cryptography==2.3     # CVE-2018-10903 - ECDH point validation
urllib3==1.24.1       # CVE-2019-11324 - Certificate verification bypass
django==2.1.0         # CVE-2019-6975 - Memory exhaustion DoS
numpy==1.15.0         # CVE-2019-6446 - Buffer overflow

# Additional vulnerable packages
paramiko==2.4.0       # CVE-2018-1000805 - Authentication bypass
lxml==4.2.0          # CVE-2018-19787 - XPath injection
simplejson==3.15.0   # CVE-2019-12872 - Integer overflow
tornado==5.0.0       # CVE-2018-1000656 - Path traversal
